requirements.txt
PyPNG==0.20220715.0
google-api-python-client==2.96.0
google-auth-oauthlib==1.0.0
fastapi==0.101.0
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.1.1
pydantic-settings==2.0.2
asyncpg==0.28.0
httpx==0.24.1
pyjwt==2.8.0
